Added functions for building circle from two other

+++ StyleCopTasks/Point.cs
@@ -6,11 +6,11 @@ namespace StyleCopTasks
 {
     class Point
     {
-        public int X { get; set; }
+        public double X { get; set; }
 
-        public int Y { get; set; }
+        public double Y { get; set; }
 
-        public Point(int x, int y)
+        public Point(double x, double y)
         {
             X = x;
             Y = y;
